requirements.txt
Flask==3.1.2
python-dateutil==2.9.0.post0
Werkzeug==3.1.5
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
gunicorn==21.2.0
